Move LineStringIterator
