+++ src/scalar/linestring/mod.rs
@@ -1,5 +1,7 @@
+pub mod iterator;
 mod owned;
 mod scalar;
 
+pub use iterator::LineStringIterator;
 pub use owned::OwnedLineString;
 pub use scalar::LineString;
